requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
Werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
redis==4.6.0